# backend-innotech/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InnoTech Preinscription API",
        "description": "Documentation for the InnoTech University APIS ",
        "contact": {
            "name": "Université InnoTech",
            "email": "admissions@InnoTech-university.tn"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/formulaires/confirm-preinscription/{application}": {
            "get": {
                "tags": [
                    "Confirmation"
                ],
                "summary": "Afficher le formulaire de confirmation tout en listant les champs deja remplis dans fromualire de préinscription et les champs pour joindre les documents selon level (Licence/Mastere) mentionné dans l'appplication",
                "operationId": "b3f338cb1e88ae3d13211eeec774392a",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID de l'application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formulaire chargé avec succès"
                    },
                    "403": {
                        "description": "Déjà soumis ou non préconfirmé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Confirmation"
                ],
                "summary": "Soumettre les documents de confirmation",
                "operationId": "263d6d398ad4255a300f02b25fb8ea53",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID de l'application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fichiers requis selon le niveau",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "documents[1]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "documents[2]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documents soumis avec succès"
                    },
                    "403": {
                        "description": "Formulaire déjà soumis ou non autorisé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/degrees": {
            "get": {
                "tags": [
                    "Degrees"
                ],
                "summary": "List all degrees",
                "operationId": "79d098a12a4303d67f4e9cd3e6adfa14",
                "responses": {
                    "200": {
                        "description": "Degrees retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Create a new degree",
                "operationId": "9d962cbad9df1fe20ebd769f9ececba7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level",
                                    "translations"
                                ],
                                "properties": {
                                    "level": {
                                        "type": "string"
                                    },
                                    "translations": {
                                        "properties": {
                                            "fr": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "en": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "ar": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Degree created successfully"
                    }
                }
            }
        },
        "/api/degrees/{id}": {
            "get": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Get a degree by ID",
                "operationId": "199bdca35978f6c8bb70e04d04a39b44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Degree retrieved successfully"
                    }
                }
            },
            "put": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Update a degree",
                "description": "Update a degree by its ID",
                "operationId": "updateDegree",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the degree to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Génie Logiciel"
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Licence"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Degree updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Degree not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Degrees"
                ],
                "summary": "Delete a degree",
                "operationId": "ed0cdd1e25b5b7d5f96444f163dd62a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Degree deleted successfully"
                    }
                }
            }
        },
...
    },
...
}